Allow INTERNAL state aggregates to participate in partial aggregation
This adds infrastructure to allow internal states of aggregate functions
to be serialized so that they can be transferred from a worker process
into the master back-end process. The master back-end process then
performs a de-serialization of the state before performing the final
aggregate stage.

This commit does not add any serialization or de-serialization
functions. These functions will arrive in a follow-on commit.