Don't attempt to export a snapshot from CREATE_REPLICATION_SLOT when in recovery
Exporting a snapshot requires us to start an xact, and we can't do that from a
server in recovery. So skip snapshot export. The client must handle syncing of
initial state via some external means like a slot on the master or manually
stopping replay from a physical copy at the same LSN.