Add a TAP test demonstrating performance issues with slot creation
If we have an xact open when we start creating a slot and then do a lot of
other transactions, DecodingContextFindStartpoint() will examine every xact
between when the old xact started and when it commits as a candidate
startpoint. For each xact it builds a historic snapshot, doing a
qsort() on the xid list. This is O(something_bad(n)) for n transactions.

The TAP test doesn't re-run with multiple pgbench scales, it just
has a tweakeable variable.