Document that perl 5.8.4 is required for TAP tests
We don't test with anything older than Perl 5.8.4 in the buildfarm
and it's from 2004, so document that the TAP tests are expected to
work with Perl 5.8.4 and explain how to get it easily.

This does not affect configure tests; we'll still compile against
an older Perl 5.8.x if there's one in the wild.