Documentation and comments fixes relating to replication position tracking
There is some confusing code in the walsender and in logical decoding that
treats a "sentPos" variable as "sent up to but not including", and treats the
"confirmed_flush" value in a replication slot as "confirmed up to but not
including".

The latter has already resulted in an off-by-one error where confirmed_flush
was assumed to mean "confirmed flushed up and including" in some places.

Additionally, some log message output for logical decoding stated that
commits "after" the specified LSN would be replayed, when in fact it is commits
beginning at or after the specified LSN that will  be replayed.

Improve comments, messages and documentation to emphasise that sentPtr and
confirmed_flush are "up to and excluding", that the pg_stat_replication fields
are the first un-sent/un-flushed/un-committed LSN, that replication functions
start at the LSN specified inclusive, etc.

Additionally, document the restart_lsn and confirmed_flush lsn in struct
ReplicationSlotPersistentData.