Fix confusion between AttrNumber and int16 in ndistinct code
The code was a bit confused whether it handles AttrNumber or just
indexes into an array of AttrNumber values. That is mostly harmless
in terms of space for the value (2B) but it's certainly confusing.

This reworks the code to int16 when the value just an index into
stakeys, and AttrNumber for actual attribute numbers (but funnily
enough those are not used at all in the code).

This modifies not only the code in mvdist.c (combination generator
and so on), but also the MVNDistinctItem definition.