PATCH: selectivity estimation using functional  dependencies
Use functional dependencies to correct selectivity estimates of
equality clauses. For now this only works with regular WHERE
conditions, not join clauses etc.

Given two equality clauses

     (a = 1) AND (b = 2)

we compute selectivity for each condition, and then combine them
using formula

     P(a=1, b=2) = P(a=1) * [degree + (1 - degree) * P(b=2)]

where 'degree' of the functional dependence (a => b) is a number
between [0,1] measuring how much the knowledge of 'a' determines
the value of 'b'. For 'degree=0' this degrades to independence,
for 'degree=1' we get perfect functional dependency.

Estimates of more than two clauses are computed recursively, so
for example

    (a = 1) AND (b = 2) AND (c = 3)

is first split into

    P(a=1, b=2, c=3) = P(a=1, b=2) * [d + (1-d) * P(c=3)]

where 'd' is degree of (a,b => c) functional dependency. And then
the first part of the estimate is computed recursively:

    P(a=1, b=2) = P(a=1) * [d + (1-d) * P(b=2)]

where 'd' is degree of (a => b) dependency.

The patch includes regression tests with functional dependencies
on several synthetic datasets (random, perfectly correlated, etc.)