Some minor cleanup work to clausesel.c
There was no point in counting the number of relations seen in the clause
list, when we just needed to know there was more than 1 or not. This saves
calling bms_num_members(), although cheap, it seems a hot enough code
path to warrant the change.

Also fixed a few wording issues in comments.

Also simplified the loop in find_strongest_dependency()

Also removed find_stats(), this was not using the standard function to get
the RelOptInfo, and seems too simple for it to be worth being a function
at all.