Simplify clausesel matching of extended stats
The previous code pretended that we sometimes would try to match to
multiple stat types by requesting all stats matching a bitmask. This was
not true. If we do that in the future then we can add that back again.

Also removed some outdated comments in choose_ext_statistics() which
must have been left over from before using the bitmapset to match the
stats up to the attnums.

This also gets rid of the stts_type_matches() function. The function was
already over simplistic and after this change would have become even more
so, so I just removed it and moved the code into where it was required.