Rework detection method of knowing when to apply ext stats
Previously we analyzed the clause list to check if we found the clauses
to reference a single relation. This did work, but it means that we'd have
done that during the join search which could become expensive for queries
with many joins.

This commit reworks this so that we pass the relation down into
clauselist_selectivity(). Most callers know about the RelOptInfo so it
saves from having to work it out again. We also have no stats for join
relations, so we can simple pass a NULL in place of the rel for those
cases.

This commit also adds a flag to clauselist_selectivity() named
tryextstats. I'm not all that sure if that's actually needed yet, maybe
it's better to just always apply for base rels with rtekind RTE_RELATION?
I'll keep the tryextstats for now, but half of me thinks that it's really
not needed and just clutters the code a bit more.

I've also noticed that we might want to consider allowing extended stats
on foreign tables. I'll have a closer look at that and perhaps will make
some changes if I decide that they should be allowed after all.