Shuffle things around and get rid of count_ext_attnums
This function was internally calling collect_ext_attnums(), which we're
doing anyway, so this was double the work for no extra gain. It's quite
easy to shuffle things around a bit and get rid of the additional call to
collect_ext_attnums() here. I also ditched the stats list as it was only
being used in one place after having moved the has_stats_of_type call up a
bit.