Add missing function prototypes
Two of these seemed to be missing, but instead of just adding the missing
ones I've copied and pasted all of the new static functions and pasted as
prototypes in the same order they're defined in the file.
mv_compatible_walker was left out due to the struct argument being defined
later in the file.