Refactor how extended selectivity estimates are applied
This is a rather large change which moves most of the code which was added
into clausesel.c into depdendencies.c. The code in dependencies.c has now
been made much more optimal, and selectivity estimations on clauses should
perform much better than previously. We now don't split the clause list
into parts. This has been simplified, and we now use a Bitmapset to mark
which clauses (by list index) which have had selectivites applied for.

I very much tried to leave things open here for more extended stats work
in the future. clauselist_selectivity() now has a generic place which
extended stats can be applied. We can control which order these are
applied in too quite easily, just by shuffling the order that the
functions are called in. Any clauses which have already been estimated
should be ignored by the estimation function.

This commit also does a round of renaming of the various helper functions
in dependencies.c. The functions which are specific to dependencies are in
there, and the more generic functions like hash_stats_of_kind() have been
moved into extended_stats.c.

I'm still not all that happy with the code which checks of the clause is
compatible with depdency estimations. It appears that a clause such as:

(a = 10) = true may not be handled the way it thinks. Although at the
moment it does not seem to misbehave.

There's also an issue with a clause such as:

WHERE a = 10 and enabled;

would not appear to work correctly. "enabled" is not a OpExpr with 2
operands, like the code requires, so dependency stats will not be applied
here if some were to exist on (a, enabled).