Fix thinko on passing the selectivity for extended stats
For some reason I mistakenly thought that the extended stats selectivity
estimation functions must be passed in the existing selectivity and
perform actions on that. I then was multiplying the return value with the
existing statistics. Doing only one of these would have been correct, but
doing both is invalid. So to save from having to pass in the selectivity
estimates to the function, let's just make it so the caller must multiply
the return value with any existing selectivity.