Remove min group size requirement in functional deps
Previously we required that we find at least 3 rows in each group
supporting a functional dependency. This is not really required, and
having this meant that functional depencies were not found on columns that
are unique or almost unique. While this was less of an issue since an equality
qual on such a column would never suffer from row underestimations, it does
seem like a good idea to get rid of this magical number, when it's not really
required.