Remove unneeded NULL test.
I'd added this NULL test as its highly lightly that this variable is NULL
and I wanted to save from having to call bms_is_member each loop. The aim
here was to save both the function call overhead, and the useless x < 0
test that we can never possibly hit in this this code.

It seems that reviewers are disagreeing with me, so I'll just backdown and
remove it. Quite possibly the difference would not be measureable anyway.
I was just being a bit Scottish with my CPU cycles.