Issue XLOG_XACT_ASSIGNMENT with wal_level=logical
The logical decoding infrastructure needs to know which top-level
transaction the subxact belongs to, in order to decode all the
changes. Until now that might be delayed until commit, due to the
caching (GPROC_MAX_CACHED_SUBXIDS), preventing features requiring
incremental decoding.

So when wal_level=logical, issue the subxact assignment records
immediately. This means didLogXid/log_unknown_top are unnecessary,
as we log the assignments immediately (when it matters).