Expand targetlist correctly
We must expand the targetlist to ensure that:

1. In UPDATE action, any attribute missing explicit SET clause can get the old
value

2. In INSERT action, default values can be populated correctly.

It's not clear what should be done to the targetlist at the Merge node itself
since the individual actions could either be INSERT/UPDATE/DELETE. But we
clearly can't leave it unexpanded as the tests show. So currently we just
expand them as we do in normal UPDATE. But this needs further investigation and
possible fixes.