Projection slots must not be changed
The projection slots for the tuples used by individual merge actions are
initialised when ExecInitModifyTable() calls ExecBuildProjectionInfo().
Since subsequent calls to ExecProject() will store the result in the slot
passed to ExecBuildProjectionInfo(), we must not lose track of the slot passed
to ExecBuildProjectionInfo().

This patch now ensures that we correctly see the current target tuple, when
UPDATE/INSERT actions are executed.