Add support to correctly fetch columns from the source and target relations
While projecting update/insert targetlists, we now setup the scan slot to
contain the original target tuple and the inner slot to contain the tuple from
the source relation/query. set_plan_refs() ensure that the targetlists are
fixed to refer to the appropriate slot. This way the expressions in the
targetlists are computed correctly.

Since the UPDATE/INSERT targetlists are managed separately, we must not run
JunkFilter on the tuples projected using these targetlists (these targetlists
don't contain any junk columns).