Handle additional quals in the WHEN clause
These quals are evaluated and the action is taken iff the quals return true.
When attached to NOT MATCHED clause, the quals can only refer to the columns
from the source table/query. When attached to MATCHED clause, the quals can
refer to both, source and target tables.

Subqueries in the quals are currently not supported.

Additional regression tests are added and they pass. We might want to fix a few
things though such as unfriendly error message when the source table is
referred in NOT MATCHED quals. Similarly, we get a mis-leading error message
such as "ERROR:  system column "xmin" reference in check constraint is
invalid" (and a similar one for using a subquery in the quals), which should be
fixed.