Handle column references correctly.
For INSERTs, we don't want to allow columns from the target relation to be used
in the WHEN quals and the targetlists. For DELETEs/UPDATEs, both target and
source relations are allowed.

Also, we must not resolve any unqualified column references to the top level,
internally created Join relation since the current implementation has no way to
extract values from the Join node. Instead, we resolve unqualified references
to either source or target relation, based on their visibility for the given
action.

Add a few more test cases for subquery handling.