Handle the case of join failure during EPQ
When a concurrent UPDATE updates a row such that the new row does not match
with the current source row at hand, the right outer join that MERGE is
performing will return the source row with target row filled with NULLs. We
must treat this case as "serialization error" since the original MATCHED case
is now turned into a NOT MATCHED case.